id need 64bit unsigned int

+++ assignment3/Node.hpp
@@ -8,18 +8,18 @@ class Path;
 class Node {
 public:
   // dummy node constructor
-  explicit Node(int id)
+  explicit Node(uint64_t id)
     : _point(0, 0)
     , _id(id)
     , _dummy(true)
   {}
-  explicit Node(SHPObject *obj, int id)
+  explicit Node(SHPObject *obj, uint64_t id)
     : _point(obj->dfXMin, obj->dfYMax)
     , _id(id)
     , _dummy(false)
   {}
 
-  int id() const { return _id; }
+  uint64_t id() const { return _id; }
   double x() const { return _point.x; }
   double y() const { return _point.y; }
   bool isDummy() const { return _dummy; }
@@ -28,6 +28,6 @@ public:
   std::vector<Path *> linkedPaths;
 private:
   PointD _point;
-  int _id;
+  uint64_t _id;
   bool _dummy;
 };
